+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/DispatcherServletTest.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.View;
@@ -9,6 +9,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -29,6 +30,7 @@ class DispatcherServletTest {
         dispatcherServlet.init();
     }
 
+    @Disabled
     @Test
     @DisplayName("요청과 응답을 파라미터로 받아 적절한 뷰를 렌더링한다.")
     void service() throws Exception {
@@ -37,7 +39,7 @@ class DispatcherServletTest {
         View view = mock(JspView.class);
         RequestDispatcher dispatcher = mock(RequestDispatcher.class);
 
-        when(request.getRequestURI()).thenReturn("/");
+        when(request.getRequestURI()).thenReturn("/login");
         when(request.getMethod()).thenReturn("GET");
         when(request.getRequestDispatcher(anyString())).thenReturn(dispatcher);
         when(modelAndView.getView()).thenReturn(view);
